+++ cognite_toolkit/cdf_tk/templates/_templates.py
@@ -20,14 +20,14 @@ from cognite_toolkit.cdf_tk.utils import validate_case_raw, validate_data_set_is
 
 from ._constants import COGNITE_MODULES, CUSTOM_MODULES, EXCL_INDEX_SUFFIX, PROC_TMPL_VARS_SUFFIX
 from ._utils import iterate_functions, iterate_modules
-from .data_classes import BuildConfigYAML, SystemConfig
+from .data_classes import BuildConfigYAML, SystemYAML
 
 
 def build_config(
     build_dir: Path,
     source_dir: Path,
     config: BuildConfigYAML,
-    system_config: SystemConfig,
+    system_config: SystemYAML,
     clean: bool = False,
     verbose: bool = False,
 ) -> None:
