+++ tests/tests_unit/utils.py
@@ -34,7 +34,10 @@ def mock_read_yaml_file(
         return load_yaml_inject_variables(filepath, variables, required_return_type)
 
     monkeypatch.setattr("cognite_toolkit.cdf_tk.utils.read_yaml_file", fake_read_yaml_file)
-    monkeypatch.setattr("cognite_toolkit.cdf_tk.templates.data_classes.read_yaml_file", fake_read_yaml_file)
+    monkeypatch.setattr("cognite_toolkit.cdf_tk.templates.data_classes._base.read_yaml_file", fake_read_yaml_file)
+    monkeypatch.setattr(
+        "cognite_toolkit.cdf_tk.templates.data_classes._project_directory.read_yaml_file", fake_read_yaml_file
+    )
     monkeypatch.setattr("cognite_toolkit.cdf.read_yaml_file", fake_read_yaml_file)
     monkeypatch.setattr("cognite_toolkit.cdf_tk.utils.load_yaml_inject_variables", fake_load_yaml_inject_variables)
     monkeypatch.setattr(
